Refactor navigation pages to use MultiBlocProvider for home screen
Switch from using a single BlocProvider to MultiBlocProvider for the home screen
to manage multiple Blocs (PopularCubit, RecommendedCubit, ReleasesCubit) more
efficiently. This change enhances the modularity and scalability of the navigation
system by allowing the home screen to depend on multiple data sources without
increasing complexity in the widget tree.

- Added imports for recommended and release movie repositories.
- Updated home widget to use MultiBlocProvider with providers for PopularCubit,
  RecommendedCubit, and ReleasesCubit.